layout change for home and about
