+++ layout/footerMenuItems.json
@@ -1,8 +1,8 @@
 [
 	"index",
 	"about",
-	"hosts",
-	"students",
+	"join",
+	"events",
 	"by-laws",
 	"letter",
 	"guidelines",
