+++ layout/navbarItems.json
@@ -1,7 +1,7 @@
 [
 	"index",
 	"about",
-	"hosts",
-	"students",
+	"join",
+	"events",
 	"newsletters"
 ]
